fix ut compilation error
